+++ probdiffeq/util/cholesky_util.py
@@ -20,7 +20,7 @@ manipulation of square root matrices.
 
 """
 
-from probdiffeq.backend import linalg, tree_util
+from probdiffeq.backend import control_flow, linalg, tree_util
 from probdiffeq.backend import numpy as np
 
 
@@ -141,3 +141,76 @@ def triu_via_qr(R, /):
     #  (or expose this option; some equivalence tests might fail
     #   if we always use a positive diagonal.)
     return linalg.qr_r(R)
+
+
+def cholesky_hilbert(n: int, K: int = 0):
+    """Compute the Cholesky factor of a Hilbert matrix.
+
+    This routine implements W. Kahan's stable recurrence (see "Hilbert Matrices",
+    Math H110 notes) to construct a Cholesky factor.
+
+    Parameters
+    ----------
+    n
+        Size of the Hilbert matrix (``n x n``).
+    K
+        Shift parameter. ``K = 0`` gives the classical Hilbert matrix.
+        Increasing ``K`` produces related matrices with entries
+        ``1 / (i + j + K - 1)``. Default is 0.
+
+    Returns
+    -------
+    Lower-triangular Cholesky factor of the Hilbert matrix.
+
+
+    Notes
+    -----
+    - Hilbert matrices are notoriously ill-conditioned; even with float64,
+      the factorization loses accuracy for moderately large ``n`` (≈15 or more).
+
+    References
+    ----------
+    W. Kahan, *Hilbert Matrices*,
+    https://people.eecs.berkeley.edu/~wkahan/MathH110/HilbMats.pdf
+    """
+    Kf = np.asarray(K)
+
+    odds = np.arange(K + 1, K + 2 * n, step=2)  # length n
+    dr = np.sqrt(odds)  # shape (n,)
+
+    f = np.ones((n,)) * (1.0 + Kf)
+
+    def f_body(idx, f):
+        prev = f[idx - 1]
+        idxf = np.asarray(idx)
+        val = (((prev / idxf) * (Kf + 2.0 * idxf)) / (Kf + idxf)) * (
+            Kf + 2.0 * idxf + 1.0
+        )
+        return f.at[idx].set(val)
+
+    f = control_flow.fori_loop(1, n, f_body, f)
+    f = 1.0 / f
+
+    U = np.eye(n)
+
+    def body_j(j_idx, U):
+        # compute column j_idx (0-based) of U using downward recurrence
+        g = U[:, j_idx]
+
+        def inner_body(k, g):
+            # k runs 0..j_idx-1, we want i = j_idx-1-k (descend j-1 .. 0)
+            i = j_idx - 1 - k
+            denom = np.asarray(j_idx - i)  # == k+1
+            factor = Kf + np.asarray(i + 1) + np.asarray(j_idx + 1)
+            newval = (g[i + 1] / denom) * factor
+            return g.at[i].set(newval)
+
+        g = control_flow.fori_loop(0, j_idx, inner_body, g)
+        return U.at[:, j_idx].set(g)
+
+    U = control_flow.fori_loop(1, n, body_j, U)
+
+    # scale columns: U = U .* (dr * f_row)
+    U = U * (dr[:, None] * f[None, :])
+
+    return np.tril(U.T)
